chore(test): use a global mock of InstantSearch.js root
