+++ package.json
@@ -106,6 +106,9 @@
     "snapshotSerializers": [
       "jest-serializer-html"
     ],
+    "moduleNameMapper": {
+      "instantsearch.js/es": "<rootDir>/__mocks__/instantsearch-es.js"
+    },
     "transformIgnorePatterns": [
       "node_modules/(?!(instantsearch\\.js|vue-json-tree))"
     ],
